Separate the struct definitions of the chainconfig from the parser
